move non-grpc code to //src/server/metadata
