+++ src/server/metadata/server/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@rules_go//go:def.bzl", "go_library", "go_test")
+load("@rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "server",
@@ -6,22 +6,10 @@ go_library(
     importpath = "github.com/pachyderm/pachyderm/v2/src/server/metadata/server",
     visibility = ["//visibility:public"],
     deps = [
-        "//src/internal/dbutil",
-        "//src/internal/errors",
         "//src/internal/pachsql",
-        "//src/internal/pfsdb",
         "//src/metadata",
-    ],
-)
-
-go_test(
-    name = "server_test",
-    size = "small",
-    srcs = ["server_test.go"],
-    embed = [":server"],
-    deps = [
-        "//src/metadata",
-        "@com_github_google_go_cmp//cmp",
-        "@com_github_google_go_cmp//cmp/cmpopts",
+        "//src/server/metadata",
+        "@org_golang_google_grpc//codes",
+        "@org_golang_google_grpc//status",
     ],
 )
